Renaming and providing meaningful names to classes and variables.
