+++ app/src/main/java/com/karansyd4/newsappexercise/data/remote/ResultResponse.kt
@@ -2,7 +2,7 @@ package com.karansyd4.newsappexercise.data.remote
 
 import com.google.gson.annotations.SerializedName
 
-data class ResultsResponse<T>(
+data class NewsResponse<T>(
     @SerializedName("rows")
     val rows: List<T>,
     @SerializedName("title")
