+++ app/src/main/java/com/karansyd4/newsappexercise/data/remote/datasource/NewsDataSource.kt
@@ -4,7 +4,7 @@ import com.karansyd4.newsappexercise.data.remote.BaseDataSource
 import com.karansyd4.newsappexercise.data.remote.NewsService
 import javax.inject.Inject
 
-class RemoteDataSource @Inject constructor(private val newsService: NewsService) :
+class NewsDataSource @Inject constructor(private val newsService: NewsService) :
     BaseDataSource() {
     suspend fun fetchNews() = getResult { newsService.fetchNews() }
 }
