+++ app/src/main/java/com/karansyd4/newsappexercise/data/remote/dto/News.kt
@@ -5,7 +5,7 @@ import com.google.gson.annotations.SerializedName
 /**
  * Defined all the variables as @var, because it might happen that backend send some values as null
  */
-data class Row(
+data class News(
     @SerializedName("description")
     var description: String? = null,
     @SerializedName("imageHref")
